Move AWS specific code to separate directory
