+++ packages/server/src/auth/setpassword.test.ts
@@ -23,6 +23,7 @@ const app = express();
 describe('Set Password', () => {
   beforeAll(async () => {
     const config = await loadTestConfig();
+    config.emailProvider = 'awsses';
     await initApp(app, config);
   });
