+++ packages/server/src/email/email.ts
@@ -1,13 +1,12 @@
-import { SendEmailCommand, SESv2Client } from '@aws-sdk/client-sesv2';
-import { badRequest, normalizeErrorString, OperationOutcomeError } from '@medplum/core';
 import { Binary } from '@medplum/fhirtypes';
 import { createTransport } from 'nodemailer';
-import MailComposer from 'nodemailer/lib/mail-composer';
-import Mail, { Address } from 'nodemailer/lib/mailer';
+import Mail from 'nodemailer/lib/mailer';
+import { sendEmailViaSes } from '../cloud/aws/email';
 import { getConfig, MedplumSmtpConfig } from '../config';
 import { Repository } from '../fhir/repo';
 import { getBinaryStorage } from '../fhir/storage';
 import { globalLogger } from '../logger';
+import { getFromAddress } from './utils';
 
 /**
  * Sends an email using the AWS SES service.
@@ -19,7 +18,6 @@ import { globalLogger } from '../logger';
 export async function sendEmail(repo: Repository, options: Mail.Options): Promise<void> {
   const config = getConfig();
   const fromAddress = getFromAddress(options);
-  const toAddresses = buildAddresses(options.to);
 
   options.from = fromAddress;
   options.sender = fromAddress;
@@ -32,83 +30,13 @@ export async function sendEmail(repo: Repository, options: Mail.Options): Promis
   // "if set to true then fails with an error when a node tries to load content from a file"
   options.disableFileAccess = true;
 
-  globalLogger.info('Sending email', { to: toAddresses?.join(', '), subject: options.subject });
+  globalLogger.info('Sending email', { to: options.to, subject: options.subject });
 
   if (config.smtp) {
     await sendEmailViaSmpt(config.smtp, options);
-  } else {
+  } else if (config.emailProvider === 'awsses') {
     await sendEmailViaSes(options);
   }
-}
-
-/**
- * Returns the from address to use.
- * If the user specified a from address, it must be an approved sender.
- * Otherwise uses the support email address.
- * @param options - The user specified nodemailer options.
- * @returns The from address to use.
- */
-function getFromAddress(options: Mail.Options): string {
-  const config = getConfig();
-
-  if (options.from) {
-    const fromAddress = addressToString(options.from);
-    if (fromAddress && config.approvedSenderEmails?.split(',')?.includes(fromAddress)) {
-      return fromAddress;
-    }
-  }
-
-  return config.supportEmail;
-}
-
-/**
- * Converts nodemailer addresses to an array of strings.
- * @param input - nodemailer address input.
- * @returns Array of string addresses.
- */
-function buildAddresses(input: string | Address | (string | Address)[] | undefined): string[] | undefined {
-  if (!input) {
-    return undefined;
-  }
-  if (Array.isArray(input)) {
-    return input.map(addressToString) as string[];
-  }
-  return [addressToString(input) as string];
-}
-
-/**
- * Converts a nodemailer address to a string.
- * @param address - nodemailer address input.
- * @returns String address.
- */
-function addressToString(address: Address | string | undefined): string | undefined {
-  if (address) {
-    if (typeof address === 'string') {
-      return address;
-    }
-    if (typeof address === 'object' && 'address' in address) {
-      return address.address;
-    }
-  }
-  return undefined;
-}
-
-/**
- * Builds a raw email message using nodemailer MailComposer.
- * @param options - The nodemailer options.
- * @returns The raw email message.
- */
-function buildRawMessage(options: Mail.Options): Promise<Uint8Array> {
-  const msg = new MailComposer(options);
-  return new Promise((resolve, reject) => {
-    msg.compile().build((err, message) => {
-      if (err) {
-        reject(err);
-        return;
-      }
-      resolve(message);
-    });
-  });
 }
 
 /**
@@ -169,40 +97,4 @@ async function sendEmailViaSmpt(smtpConfig: MedplumSmtpConfig, options: Mail.Opt
     },
   });
   await transport.sendMail(options);
-}
-
-/**
- * Sends an email via AWS SES.
- * @param options - The nodemailer options.
- */
-async function sendEmailViaSes(options: Mail.Options): Promise<void> {
-  const config = getConfig();
-  const fromAddress = addressToString(options.from);
-  const toAddresses = buildAddresses(options.to);
-  const ccAddresses = buildAddresses(options.cc);
-  const bccAddresses = buildAddresses(options.bcc);
-
-  let msg: Uint8Array;
-  try {
-    msg = await buildRawMessage(options);
-  } catch (err) {
-    throw new OperationOutcomeError(badRequest('Invalid email options: ' + normalizeErrorString(err)), err);
-  }
-
-  const sesClient = new SESv2Client({ region: config.awsRegion });
-  await sesClient.send(
-    new SendEmailCommand({
-      FromEmailAddress: fromAddress,
-      Destination: {
-        ToAddresses: toAddresses,
-        CcAddresses: ccAddresses,
-        BccAddresses: bccAddresses,
-      },
-      Content: {
-        Raw: {
-          Data: msg,
-        },
-      },
-    })
-  );
 }
