+++ packages/server/src/fhir/operations/execute.ts
@@ -1,4 +1,3 @@
-import { InvokeCommand, LambdaClient } from '@aws-sdk/client-lambda';
 import {
   ContentType,
   Hl7Message,
@@ -7,7 +6,6 @@ import {
   allOk,
   badRequest,
   createReference,
-  getIdentifier,
   normalizeErrorString,
   resolveId,
 } from '@medplum/core';
@@ -30,8 +28,8 @@ import fetch from 'node-fetch';
 import { randomUUID } from 'node:crypto';
 import { Readable } from 'node:stream';
 import vm from 'node:vm';
-import { TextDecoder, TextEncoder } from 'util';
 import { asyncWrap } from '../../async';
+import { runInLambda } from '../../cloud/aws/execute';
 import { getConfig } from '../../config';
 import { buildTracingExtension, getAuthenticatedContext, getLogger } from '../../context';
 import { generateAccessToken } from '../../oauth/keys';
@@ -57,6 +55,11 @@ export interface BotExecutionRequest {
   readonly forwardedFor?: string;
   readonly requestTime?: string;
   readonly traceId?: string;
+}
+
+export interface BotExecutionContext extends BotExecutionRequest {
+  readonly accessToken: string;
+  readonly secrets: Record<string, ProjectSecret>;
 }
 
 export interface BotExecutionResult {
@@ -161,7 +164,7 @@ async function getBotForRequest(req: Request): Promise<Bot | undefined> {
  * @returns The bot execution result.
  */
 export async function executeBot(request: BotExecutionRequest): Promise<BotExecutionResult> {
-  const { bot } = request;
+  const { bot, runAs } = request;
   const startTime = request.requestTime ?? new Date().toISOString();
 
   let result: BotExecutionResult;
@@ -172,10 +175,16 @@ export async function executeBot(request: BotExecutionRequest): Promise<BotExecu
   } else {
     await writeBotInputToStorage(request);
 
+    const context: BotExecutionContext = {
+      ...request,
+      accessToken: await getBotAccessToken(runAs),
+      secrets: await getBotSecrets(bot),
+    };
+
     if (bot.runtimeVersion === 'awslambda') {
-      result = await runInLambda(request);
+      result = await runInLambda(context);
     } else if (bot.runtimeVersion === 'vmcontext') {
-      result = await runInVmContext(request);
+      result = await runInVmContext(context);
     } else {
       result = { success: false, logResult: 'Unsupported bot runtime' };
     }
@@ -276,117 +285,12 @@ async function writeBotInputToStorage(request: BotExecutionRequest): Promise<voi
 }
 
 /**
- * Executes a Bot in an AWS Lambda.
- * @param request - The bot request.
- * @returns The bot execution result.
- */
-async function runInLambda(request: BotExecutionRequest): Promise<BotExecutionResult> {
-  const { bot, runAs, input, contentType, traceId } = request;
-  const config = getConfig();
-  const accessToken = await getBotAccessToken(runAs);
-  const secrets = await getBotSecrets(bot);
-
-  const client = new LambdaClient({ region: config.awsRegion });
-  const name = getLambdaFunctionName(bot);
-  const payload = {
-    bot: createReference(bot),
-    baseUrl: config.baseUrl,
-    accessToken,
-    input: input instanceof Hl7Message ? input.toString() : input,
-    contentType,
-    secrets,
-    traceId,
-  };
-
-  // Build the command
-  const encoder = new TextEncoder();
-  const command = new InvokeCommand({
-    FunctionName: name,
-    InvocationType: 'RequestResponse',
-    LogType: 'Tail',
-    Payload: encoder.encode(JSON.stringify(payload)),
-  });
-
-  // Execute the command
-  try {
-    const response = await client.send(command);
-    const responseStr = response.Payload ? new TextDecoder().decode(response.Payload) : undefined;
-
-    // The response from AWS Lambda is always JSON, even if the function returns a string
-    // Therefore we always use JSON.parse to get the return value
-    // See: https://stackoverflow.com/a/49951946/2051724
-    const returnValue = responseStr ? JSON.parse(responseStr) : undefined;
-
-    return {
-      success: !response.FunctionError,
-      logResult: parseLambdaLog(response.LogResult as string),
-      returnValue,
-    };
-  } catch (err) {
-    return {
-      success: false,
-      logResult: normalizeErrorString(err),
-    };
-  }
-}
-
-/**
- * Returns the AWS Lambda function name for the given bot.
- * By default, the function name is based on the bot ID.
- * If the bot has a custom function, and the server allows it, then that is used instead.
- * @param bot - The Bot resource.
- * @returns The AWS Lambda function name.
- */
-export function getLambdaFunctionName(bot: Bot): string {
-  if (getConfig().botCustomFunctionsEnabled) {
-    const customFunction = getIdentifier(bot, 'https://medplum.com/bot-external-function-id');
-    if (customFunction) {
-      return customFunction;
-    }
-  }
-
-  // By default, use the bot ID as the Lambda function name
-  return `medplum-bot-lambda-${bot.id}`;
-}
-
-/**
- * Parses the AWS Lambda log result.
- *
- * The raw logs include markup metadata such as timestamps and billing information.
- *
- * We only want to include the actual log contents in the AuditEvent,
- * so we attempt to scrub away all of that extra metadata.
- *
- * See: https://docs.aws.amazon.com/lambda/latest/dg/nodejs-logging.html
- * @param logResult - The raw log result from the AWS lambda event.
- * @returns The parsed log result.
- */
-function parseLambdaLog(logResult: string): string {
-  const logBuffer = Buffer.from(logResult, 'base64');
-  const log = logBuffer.toString('ascii');
-  const lines = log.split('\n');
-  const result = [];
-  for (const line of lines) {
-    if (line.startsWith('START RequestId: ')) {
-      // Ignore start line
-      continue;
-    }
-    if (line.startsWith('END RequestId: ') || line.startsWith('REPORT RequestId: ')) {
-      // Stop at end lines
-      break;
-    }
-    result.push(line);
-  }
-  return result.join('\n').trim();
-}
-
-/**
  * Executes a Bot on the server in a separate Node.js VM.
  * @param request - The bot request.
  * @returns The bot execution result.
  */
-async function runInVmContext(request: BotExecutionRequest): Promise<BotExecutionResult> {
-  const { bot, runAs, input, contentType, traceId } = request;
+async function runInVmContext(request: BotExecutionContext): Promise<BotExecutionResult> {
+  const { bot, input, contentType, traceId } = request;
 
   const config = getConfig();
   if (!config.vmContextBotsEnabled) {
@@ -405,9 +309,6 @@ async function runInVmContext(request: BotExecutionRequest): Promise<BotExecutio
   const binary = await systemRepo.readReference<Binary>({ reference: codeUrl } as Reference<Binary>);
   const stream = await getBinaryStorage().readBinary(binary);
   const code = await readStreamToString(stream);
-
-  const accessToken = await getBotAccessToken(runAs);
-  const secrets = await getBotSecrets(bot);
   const botConsole = new MockConsole();
 
   const sandbox = {
@@ -420,10 +321,10 @@ async function runInVmContext(request: BotExecutionRequest): Promise<BotExecutio
     event: {
       bot: createReference(bot),
       baseUrl: config.baseUrl,
-      accessToken,
+      accessToken: request.accessToken,
       input: input instanceof Hl7Message ? input.toString() : input,
       contentType,
-      secrets,
+      secrets: request.secrets,
       traceId,
     },
   };
